refactor image loader component and mode it to design system
