+++ client_end_user/shared/src/commonMain/kotlin/presentation/composable/PreferredFoodCard.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
+import com.beepbeep.designSystem.ui.composable.BpImageLoader
 import com.beepbeep.designSystem.ui.theme.Theme
 import presentation.cuisines.CuisineUiState
 import resources.Resources
