+++ client_end_user/shared/src/commonMain/kotlin/presentation/taxi/TaxiOrderScreen.kt
@@ -127,5 +127,4 @@ class TaxiOrderScreen :
             )
         }
     }
-
 }
